+++ CHANGELOG.md
@@ -20,7 +20,9 @@
 - Images are now saved with their proper extensions.
 - Limited image format support to bmp, jpeg, png, tiff, gif, and ico, ensuring compatibility with WPF.
 - Removed Unused Image Garbabe Collector Configurables (`IMAGE_GC_INTERVAL_MINUTES` & `IMAGE_GC_KEEP_DAYS`)
-- Implemented Serverside Bandwith Limit Configuration `SERVER_MAX_DOWNLOAD_BANDWIDTH_IN_KBPS` to control the servers max bandwidth #10
+- Implemented Serverside Bandwidth Limit Configuration `SERVER_MAX_DOWNLOAD_BANDWIDTH_IN_KBPS` to control the servers max bandwidth #10
+- Implemented Clientside Bandwidth Limit Configuration by `max_kbps` Header in Download Request
+- Admins can now delete other Users Progresses by the Progress ID
 
 ## 4.0.1